Do some type hinting while understanding code paths
