+++ realkd/loss.py
@@ -1,0 +1,137 @@
+from abc import ABC, abstractmethod
+from numpy import  exp, floating, full_like, log2, stack, zeros_like
+from numpy.typing import ArrayLike, NDArray
+
+class AbsLoss(ABC):
+    @staticmethod
+    @abstractmethod
+    def __call__(y: ArrayLike, s: ArrayLike) -> float: pass
+
+    @staticmethod
+    @abstractmethod
+    def predictions(s) -> NDArray: pass
+
+    @staticmethod
+    @abstractmethod
+    def g(y, s) -> NDArray[floating]: pass
+
+    @staticmethod
+    @abstractmethod
+    def h(y, s) -> NDArray[floating]: pass
+
+    @staticmethod
+    @abstractmethod
+    def __repr__() -> str: pass
+
+    @staticmethod
+    @abstractmethod
+    def __str__() -> str: pass
+
+class SquaredLoss(AbsLoss):
+    """
+    Squared loss function l(y, s) = (y-s)^2.
+
+    >>> squared_loss
+    squared_loss
+    >>> y = array([-2, 0, 3])
+    >>> s = array([0, 1, 2])
+    >>> squared_loss(y, s)
+    array([4, 1, 1])
+    >>> squared_loss.g(y, s)
+    array([ 4,  2, -2])
+    >>> squared_loss.h(y, s)
+    array([2, 2, 2])
+    """
+
+    _instance = None
+
+    def __new__(cls):
+        if cls._instance is None:
+            cls._instance = super(SquaredLoss, cls).__new__(cls)
+        return cls._instance
+
+    @staticmethod
+    def __call__(y, s):
+        return (y - s)**2  # type: ignore 
+
+    @staticmethod
+    def predictions(s):
+        return s
+
+    @staticmethod
+    def g(y, s):
+        return -2*(y - s)
+
+    @staticmethod
+    def h(y, s):
+        return full_like(s, 2)  # Series(full_like(s, 2))
+
+    @staticmethod
+    def __repr__():
+        return 'squared_loss'
+
+    @staticmethod
+    def __str__():
+        return 'squared'
+
+
+class LogisticLoss(AbsLoss):
+    """
+    Logistic loss function l(y, s) = log2(1 + exp(-ys)).
+
+    Function assumes that positive and negative values are encoded as +1 and -1, respectively.
+
+    >>> y = array([1, -1, 1, -1])
+    >>> s = array([0, 0, 10, 10])
+    >>> logistic_loss(y, s)
+    array([1.00000000e+00, 1.00000000e+00, 6.54967668e-05, 1.44270159e+01])
+    >>> logistic_loss.g(y, s)
+    array([-5.00000000e-01,  5.00000000e-01, -4.53978687e-05,  9.99954602e-01])
+    >>> logistic_loss.h(y, s)
+    array([2.50000000e-01, 2.50000000e-01, 4.53958077e-05, 4.53958077e-05])
+    """
+
+    _instance = None
+
+    def __new__(cls):
+        if cls._instance is None:
+            cls._instance = super(LogisticLoss, cls).__new__(cls)
+        return cls._instance
+
+    @staticmethod
+    def __call__(y, s):
+        return log2(1 + exp(-y*s)) # type: ignore 
+
+    @staticmethod
+    def sigmoid(a):
+        return 1 / (1 + exp(-a))
+
+    @staticmethod
+    def predictions(s):
+        preds = zeros_like(s)
+        preds[s >= 0] = 1
+        preds[s < 0] = -1
+        return preds  # this case now returns np array
+
+    @staticmethod
+    def probabilities(s):
+        pos = LogisticLoss.sigmoid(s)
+        return stack((1-pos, pos), axis=1)
+
+    @staticmethod
+    def g(y, s):
+        return -y*LogisticLoss.sigmoid(-y*s)
+
+    @staticmethod
+    def h(y, s):
+        sig = LogisticLoss.sigmoid(-y*s)
+        return sig*(1.0-sig)
+
+    @staticmethod
+    def __repr__():
+        return 'logistic_loss'
+
+    @staticmethod
+    def __str__():
+        return 'logistic'
+
